Reg move v2
