+++ Common/src/main/java/potionstudios/byg/common/world/feature/stateproviders/BYGStateProviders.java
@@ -6,8 +6,8 @@ import net.minecraft.world.level.levelgen.feature.stateproviders.BlockStateProvi
 import net.minecraft.world.level.levelgen.feature.stateproviders.BlockStateProviderType;
 import potionstudios.byg.BYG;
 import potionstudios.byg.mixin.access.BlockStateProviderAccess;
-import potionstudios.byg.registration.RegistrationProvider;
-import potionstudios.byg.registration.RegistryObject;
+import potionstudios.byg.reg.RegistrationProvider;
+import potionstudios.byg.reg.RegistryObject;
 
 public class BYGStateProviders {
     private static final RegistrationProvider<BlockStateProviderType<?>> PROVIDER = RegistrationProvider.get(Registry.BLOCK_STATE_PROVIDER_TYPE_REGISTRY, BYG.MOD_ID);
