+++ Common/src/main/java/potionstudios/byg/reg/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * So, if you got here, most probably you want to see the registration system of this mod.
+ * Well, you are in the wrong place. This mod uses <a href="https://github.com/Matyrobbrt/RegistrationUtils">RegistrationUtils</a> for
+ * all its registration needs.
+ */
+package potionstudios.byg.reg;
