+++ build.gradle
@@ -1,10 +1,10 @@
 plugins {
     id 'java'
-    id 'com.matyrobbrt.mc.registrationutils' version '0.2.0'
+    id 'com.matyrobbrt.mc.registrationutils' version '0.2.2'
 }
 
 registrationUtils {
-    group 'potionstudios.byg.registration'
+    group 'potionstudios.byg.reg'
     projects {
         Common { type 'common'; project ':Common' }
         Fabric { type 'fabric'; project ':Fabric' }
